Leetcode 540
Here I learnt:
When you want to cancel out duplicate elements always use XOR.
Whenever you are using a vector to get input either define the size or get size from the user never leave it as undeclared.
Never iterate the array and simultaneously change it's size. It will lead you to an infinite loop.
Try to pass vector as pass by reference when you want to interpolate or operate directly on vector or array.